data-loader/pyproject.toml
[tool.poetry]
name = "data-loader"
version = "0.1.0"
description = ""
authors = ["Fco Aguilera <aguimorefran@gmail.com>"]
readme = "README.md"
packages = [{include = "data_loader"}]

[tool.poetry.dependencies]
python = "^3.7"
cassandra-driver = "^3.25.0"
pandas = ">1.1.0"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
